@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "beets"
-version = "2.0.0"
+version = "2.2.0"
 description = "music tagger and library organizer"
 authors = ["Adrian Sampson <adrian@radbox.org>"]
 maintainers = ["Serene-Arc"]
@@ -17,7 +17,6 @@ classifiers = [
     "Environment :: Web Environment",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: Implementation :: CPython",
@@ -26,13 +25,21 @@ packages = [
     { include = "beets" },
     { include = "beetsplug" },
 ]
+include = [ # extra files to include in the sdist
+    "docs",
+    "extra",
+    "man/**/*",
+    "test/*.py",
+    "test/rsrc/**/*",
+]
+exclude = ["docs/_build", "docs/modd.conf", "docs/**/*.css"]
 
 [tool.poetry.urls]
 Changelog = "https://github.com/beetbox/beets/blob/master/docs/changelog.rst"
 "Bug Tracker" = "https://github.com/beetbox/beets/issues"
 
 [tool.poetry.dependencies]
-python = ">=3.8,<4"
+python = ">=3.9,<4"
 
 colorama = { version = "*", markers = "sys_platform == 'win32'" }
 confuse = ">=1.5.0"
@@ -149,6 +156,15 @@ build-backend = "poetry.core.masonry.api"
 poethepoet = ">=0.26"
 poetry = ">=1.8"
 
+[tool.poe.tasks.build]
+help = "Build the package"
+shell = """
+make -C docs man
+rm -rf man
+mv docs/_build/man .
+poetry build
+"""
+
 [tool.poe.tasks.bump]
 help = "Bump project version and update relevant files"
 cmd = "python ./extra/release.py bump $version"
@@ -224,7 +240,7 @@ done
 interpreter = "zsh"
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 line-length = 80
 
 [tool.ruff.lint]
@@ -239,6 +255,7 @@ select = [
     "PT", # flake8-pytest-style
     # "RUF", # ruff
     # "UP", # pyupgrade
+    "TCH", # flake8-type-checking
     "W", # pycodestyle
 ]
 [tool.ruff.lint.per-file-ignores]
